script/verify_flags: extend script_verify_flags to 64 bits

+++ src/script/verify_flags.h
@@ -9,12 +9,12 @@
 #include <compare>
 #include <cstdint>
 
-enum class script_verify_flag_name : uint32_t;
+enum class script_verify_flag_name : uint64_t;
 
 class script_verify_flags
 {
 public:
-    using value_type = uint32_t;
+    using value_type = uint64_t;
 
     consteval script_verify_flags() = default;
